Fix author validation in Delete
